requirements-dev.txt
fabric
PyYAML
mock
nose
flake8
flask
